requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.6
colorama==0.4.5
cryptography==41.0.5
cssselect2==0.7.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.19
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.3.2
frozenlist==1.4.0
greenlet==2.0.2
html5lib==1.1
idna==3.4
inflect==7.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.2.0
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.4.12
numpy==1.23.4
openai==0.27.8
oscrypto==1.3.0
packaging==21.3
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.13.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pyparsing==3.0.9
pypdf==3.16.4
PyPDF2==3.0.1
pypiwin32==223
pypng==0.20220715.0
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
PythonTurtle==0.3.2
pytz==2022.4
pywin32==304
pywinpty==2.0.8
PyYAML==6.0.1
pyzmq==24.0.1
qrcode==7.4.2
qtconsole==5.3.2
QtPy==2.2.1
reportlab==3.6.13
requests==2.28.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.15
sqlparse==0.4.3
stack-data==0.5.1
svglib==1.5.1
tabula==1.0.5
tabula-py==2.6.0
terminado==0.16.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
typing_extensions==4.6.3
tzdata==2022.5
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.13
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.6
widgetsnbextension==4.0.3
xhtml2pdf==0.2.11
yarl==1.9.2
